% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(resA, resB)
}
\arguments{
\item{intercept}{}
}
\value{
result
}
\description{

}
\details{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan(rate=0, fp_error_rate=1, fn_error_rate=1)
simB <- simulate_pan(rate=0, fp_error_rate=5, fn_error_rate=5)
resA <- fit_tweedie(simA$pa, simA$tree, nboot=0)
resB <- fit_tweedie(simB$pa, simB$tree, nboot=0)
plot_pangenome_fits(list(a=resA, b=resB))
comp <- compare_pangenomes(resA, resB)

}
