% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tweedie.R
\name{fit_tweedie}
\alias{fit_tweedie}
\title{fit_tweedie}
\usage{
fit_tweedie(
  pa,
  tree,
  nboot = 100,
  mintip = 0.001,
  quiet = FALSE,
  stochastic.mapping = TRUE
)
}
\arguments{
\item{pa}{a binary gene presence/absence matrix with genes as columns and genomes as rows. The rownames should match the tip.labels of the corresponding phylogeny.}

\item{tree}{a core gene phylogney of class \link{phylo}}

\item{nboot}{the number of bootstrap replicates to perform (default=100)}

\item{quiet}{whether to print progress information (default=FALSE)}
}
\value{
result
}
\description{

}
\details{
Fits a tweedie...
}
\examples{

sim <- simulate_pan(rate=1e-3,mean_trans_size=3)
pa <- sim$pa
tree <- sim$tree
nboot <- 100
res <- fit_tweedie(sim$pa, sim$tree)


}
