% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panplotter-package.R
\docType{package}
\name{panplotter-package}
\alias{panplotter}
\alias{panplotter-package}
\title{panplotter: Pangenome Plots}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\keyword{internal}
