% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acc.R
\name{plot_acc}
\alias{plot_acc}
\title{plot_acc}
\usage{
plot_acc(pa, nperm = 100, legend = TRUE, text_size = 14, color_pallete = 6)
}
\arguments{
\item{pa}{a gene precence/absence matrix. Alternatively a list of matrices can be given.}

\item{text_size}{adjusts the size of text in the plot}

\item{color_pallete}{the colour pallete to use. A number between 1 & 9. See 'scale_colour_brewer' for more details}
}
\value{
a plot of the pangenome accumulation curve
}
\description{

}
\details{
Plots a pangenome accumulation curve.
}
\examples{

sim <- simulate_pan(rate=1e-3)
pa <- sim$pa
plot_acc(pa, color_pallete=7)

}
