% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_params.R
\name{plot_dist_params}
\alias{plot_dist_params}
\title{plot_dist_params}
\usage{
plot_dist_params(fit, plot = TRUE)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple results can be passed as a named list.}

\item{plot}{(default=FALSE)}
}
\value{
result
}
\description{

}
\details{
Plots the fitted pangenome tweedie regression model.
}
\examples{

sim <- simulate_pan(rate=1e-3)
fA <- panstripe(sim$pa, sim$tree, nboot=10)
plot_dist_params(fA)
sim <- simulate_pan(rate=1e-2)
fB <- panstripe(sim$pa, sim$tree, nboot=10)
plot_dist_params(list(a=fA,b=fB))

}
