% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise.R
\name{plot_pairwise}
\alias{plot_pairwise}
\title{plot_pairwise}
\usage{
plot_pairwise(
  fit,
  plot = TRUE,
  smooth = FALSE,
  legend = TRUE,
  text_size = 14,
  color_pallete = 6,
  facet = FALSE
)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple fits can be passed as a named list.}

\item{plot}{whether to generate the plot (default) or return a data.frame}

\item{smooth}{adds a linear line to aid the eye in seeing patters. This is NOT the model fit inferred by Panstripe. (default=FALSE)}

\item{legend}{toggles the display of the legend on and off}

\item{text_size}{the base text size of the plot (default=14)}

\item{color_pallete}{the pallete number passed to `scale_fill_brewer`}

\item{facet}{whether or not to generate separate plots for each pangenome (default=FALSE).}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the pairwise cumulative branch length versus the cumulative number of gene gain and loss events.
}
\examples{

sim <- simulate_pan(rate=1e-4)
fA <- panstripe(sim$pa, sim$tree, nboot=0)
plot_pangenome_cumulative(fA, color_pallete=6)
sim <- simulate_pan(rate=1e-3)
fB <-panstripe(sim$pa, sim$tree, nboot=0)
fit <- fB
plot=TRUE
legend=TRUE
text_size=14
color_pallete=6
facet=FALSE
plot_pairwise(list(a=fA,b=fB), color_pallete=6, smooth=TRUE)

}
