% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pangenome_cumulative.R
\name{plot_pangenome_cumulative}
\alias{plot_pangenome_cumulative}
\title{plot_pangenome_cumulative}
\usage{
plot_pangenome_cumulative(
  fit,
  plot = TRUE,
  legend = TRUE,
  text_size = 14,
  color_pallete = 6,
  facet = FALSE
)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple fits can be passed as a named list.}

\item{plot}{whether to generate the plot (default) or return a data.frame}

\item{legend}{toggles the display of the legend on and off}

\item{text_size}{the base text size of the plot (default=14)}

\item{color_pallete}{the pallete number passed to `scale_fill_brewer`}

\item{facet}{whether or not to generate separate plots for each pangenome (default=FALSE).}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the cumulative branch length versus the cumulative number of gene gain and loss events. 
This is similar to the style of plot produced by TempEst. This function plots only the data points as the fit 
can not be included as Panstripe models the individual branches rather than the cumulative total.
}
\examples{

sim <- simulate_pan(rate=0)
fA <- panstripe(sim$pa, sim$tree, nboot=0)
plot_pangenome_cumulative(fA, color_pallete=6)
sim <- simulate_pan(rate=1e-4)
fB <-panstripe(sim$pa, sim$tree, nboot=0)
plot_pangenome_cumulative(list(a=fA,b=fB), color_pallete=6)

}
