% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pangenome_params.R
\name{plot_pangenome_params}
\alias{plot_pangenome_params}
\title{plot_pangenome_params}
\usage{
plot_pangenome_params(
  fit,
  boot_ci = TRUE,
  plot = TRUE,
  legend = TRUE,
  text_size = 14,
  color_pallete = 6
)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple fits can be passed as a named list.}

\item{boot_ci}{whether to use the estimated Bootstrap confidence intervals}

\item{plot}{whether to generate the plot (default) or return a data.frame}

\item{legend}{toggles the display of the legend on and off}

\item{text_size}{the base text size of the plot (default=14)}

\item{color_pallete}{the pallete number passed to `scale_fill_brewer`}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the estimated coefficients of the pangenome regression model(s). 
It is recommended that the \link{compare_pangenomes} function is used to formally compare the slopes of different pangenome datasets.
}
\examples{

sim <- simulate_pan(rate=1e-3)
fA <- panstripe(sim$pa, sim$tree, ci_type='perc')
plot_pangenome_params(fA, color_pallete=6)
sim <- simulate_pan(rate=1e-2)
fB <-panstripe(sim$pa, sim$tree, ci_type='perc')
plot_pangenome_params(list(a=fA,b=fB), color_pallete=6, boot_ci=TRUE)

}
