% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(fit, plot = TRUE, text_size = 14)
}
\arguments{
\item{fit}{the result of running the `panstripe` function.}

\item{plot}{whether to generate the plot (default) or return a data.frame}

\item{text_size}{the base text size of the plot (default=14)}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the randomized quantile residuals of the fitted regression model as described in Dunn and Smyth (1996). This plot can be used to assess the fit of the regression.
}
\examples{

sim <- simulate_pan(rate=1e-4)
fA <- panstripe(sim$pa, sim$tree, nboot=0)
plot_residuals(fA)

}
