% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_pa.R
\name{plot_tree_pa}
\alias{plot_tree_pa}
\title{plot_tree_pa}
\usage{
plot_tree_pa(
  tree,
  pa,
  genes = colnames(pa),
  align = TRUE,
  order = TRUE,
  plot_titles = c("phylogeny", "gene presence/absence"),
  text_size = 14,
  label_genes = TRUE,
  cols = NULL
)
}
\arguments{
\item{tree}{a phylogeny in 'phylo' format}

\item{pa}{binary presence/absence matrix}

\item{genes}{a vector of gene to include in the plot. Must be a subset of colnames(pa)}

\item{align}{whether or not to align the tips of the phylogeny with the presence/absence matrix using dotted lines}

\item{order}{whether to order the genes based upon their presence/absence pattern}

\item{plot_titles}{the subplot titles (default=c('phylogeny', 'gene presence/absence'))}

\item{text_size}{adjusts the size of text in the plot}

\item{label_genes}{whether to print the gene labels in the presence/absence matrix}

\item{cols}{colours for the gene presence/absence matrix. These will be recycled if #genes > #colours.}
}
\value{
a patchwork object plotting the phylogeny alongside the gene presence/absence matrix
}
\description{
Plots a phylogeny alongside a presence/absence matrix of user selected genes.
}
\examples{

sim <- simulate_pan(rate=1e-3)
genes <- colnames(sim$pa)[which(apply(sim$pa, 2, sd)>0.2)]
plot_tree_pa(sim$tree, sim$pa, genes=genes, label_genes=FALSE, cols='black')

}
