% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsne.R
\name{plot_tsne}
\alias{plot_tsne}
\title{plot_tsne}
\usage{
plot_tsne(pa, category = NULL, plot = TRUE, perplexity = 20, pcadims = 50)
}
\arguments{
\item{pa}{binary presence/absence matrix}

\item{plot}{whether to return the plot or just the data.frame used for plotting (default=FALSE)}
}
\value{
result
}
\description{

}
\details{
Generates a tSNE plot from a gene presence/absence matrix
}
\examples{

sim <- simulate_pan(rate=1e-3)
plot_tsne(sim$pa)

}
