% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{plot_umap}
\usage{
plot_umap(pa, category = NULL, plot = TRUE, n_neighbours = 15, dim = 2, n_epochs = 200, metric = "euclidean")
}
\arguments{
\item{pa}{binary presence/absence matrix}

\item{category}{a factor vector which can be used to colour the points}

\item{plot}{whether to return the plot or just the data.frame used for plotting (default=FALSE)}

\item{perplexity}{the `perplexity` argument passed to tSNE}

\item{pcadims}{the number of principle components passed from the inital stage used in the main tSNE algorithm}

\item{n_neighbours}{number of neighbours with which the UMAP embedding is constructed}

\item{dim}{dimensionality of the target output space}

\item{n_epochs}{ number of iterations performed during layout optimisation}

\item{metric}{distance metric to determine distance between data points (see `umap.defaults` documentation for options)}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Generates a UMAP plot from a gene presence/absence matrix
}
\examples{

sim <- simulate_pan(rate=1e-3)
plot_umap(sim$pa)

}