% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfreemap.R
\name{sfreemap}
\alias{sfreemap}
\title{sfreemap}
\usage{
sfreemap(tree, pa, model = "ER", quiet = FALSE)
}
\arguments{
\item{tree}{phylo}

\item{pa}{matrix}

\item{model}{string}

\item{quiet}{logical}
}
\value{
list
}
\description{
This function has been adapted and simplified from the code in the sfreemap R package.
}
\details{
Simulation free stochastic character mapping on a phylogenetic tree.
}
\examples{

tree <- ape::rtree(10)
pa <- matrix(rbinom(10, 1, 0.5), ncol=1)
rownames(pa) <- tree$tip.label
panstripe:::sfreemap(tree, pa)

}
