% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pan.R
\name{simulate_pan}
\alias{simulate_pan}
\title{simulate_pan}
\usage{
simulate_pan(
  ngenomes = 50,
  rate = 1e-05,
  genome_length = 2000,
  fn_error_rate = 1,
  fp_error_rate = 3,
  mean_trans_size = 3,
  Q = 1,
  bf = c(0.3, 0.2)
)
}
\arguments{
\item{ngenomes}{number of genomes to simulate}

\item{rate}{rate of gene gain/loss events}

\item{genome_length}{number of genes}

\item{fn_error_rate}{false negative error rate}

\item{fp_error_rate}{false positive error rate}

\item{mean_trans_size}{mean size of the number of genes gained or lost per event}

\item{Q}{the rate matrix (default=1)}

\item{bf}{the base frequencies (default=c(0.3,0.2))}
}
\value{
A list comprising of a binary gene presence/absence matrix and a core phylogeny
}
\description{
Simulates a pangenome presence/absence matrix and accompanying core phylogeny
}
\examples{

pan <- simulate_pan()


}
