% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panfit.R
\name{validate_panfit}
\alias{validate_panfit}
\title{validate_panfit}
\usage{
validate_panfit(x)
}
\arguments{
\item{x}{panfit}
}
\description{
`validate_panfit()` checks a tibble for internal consistency.
Correct behavior can be guaranteed only if this function
runs without raising an error.
}
