% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(
  fitA,
  fitB,
  family = "Tweedie",
  ci_type = "norm",
  conf = 0.95,
  nboot = 100,
  boot_pvalue = FALSE
)
}
\arguments{
\item{fitA}{a `panfit` object generated by the `panstripe` function}

\item{fitB}{a `panfit` object generated by the `panstripe` function}

\item{family}{the family used by glm. One of 'Tweedie', 'Poisson', 'Gamma' or 'Gaussian'. (default='Tweedie')}

\item{ci_type}{the method used to calculate the bootstrap CI (default='norm'). See \link[boot]{boot.ci} for more details.}

\item{conf}{A scalar indicating the confidence level of the required intervals (default=0.95)}

\item{nboot}{the number of bootstrap replicates to perform (default=100)}

\item{boot_pvalue}{whether or not to calculate bootstrap p-values (default=FALSE)}
}
\value{
a list containing a summary of the comparison and the resulting `glm` model object
}
\description{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan(rate=1e-3, ngenomes = 100, fn_error_rate=1, fp_error_rate=1, mean_trans_size=3)
simB <- simulate_pan(rate=1e-3, ngenomes = 100, fn_error_rate=1, fp_error_rate=1, mean_trans_size=6)
fitA <- panstripe(simA$pa, simA$tree, nboot=100, ci_type='norm')
fitA$summary
fitB <- panstripe(simB$pa, simB$tree, nboot=100, ci_type='norm')
fitB$summary
comp <- compare_pangenomes(fitA, fitB, ci_type='norm', boot_pvalue=TRUE)
comp$summary
plot_dist_params(list(L=fitA, S=fitB), boot_ci=TRUE)

}
