% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(fitA, fitB, family = "Tweedie", test = "F")
}
\arguments{
\item{fitA}{a `panfit` object generated by the `panstripe` function}

\item{fitB}{a `panfit` object generated by the `panstripe` function}

\item{family}{the family used by glm. One of 'Tweedie', 'Poisson', 'Gamma' or 'Gaussian'. (default='Tweedie')}

\item{test}{the test to be performed by anova (default='F')}
}
\value{
a list containing a summary of the comparison and the resulting `glm` model object
}
\description{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan(rate=1e-3, ngenomes = 200, fn_error_rate=1, fp_error_rate=1)
simB <- simulate_pan(rate=1e-3, ngenomes = 200, fn_error_rate=1, fp_error_rate=1)
fitA <- panstripe(simA$pa, simA$tree, boot_type='norm')
fitB <- panstripe(simB$pa, simB$tree, boot_type='norm')
comp <- compare_pangenomes(fitA, fitB)

}
