% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(fitA, fitB)
}
\arguments{
\item{fitA}{a `panfit` object generated by the `panstripe` function}

\item{fitB}{a `panfit` object generated by the `panstripe` function}
}
\value{
a list containing a summary of the comparison and the resulting `glm` model object
}
\description{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan(rate=1e-3)
simB <- simulate_pan(rate=1e-4)
fA <- panstripe(simA$pa, simA$tree, nboot=0)
fB <- panstripe(simB$pa, simB$tree, nboot=0)
comp <- compare_pangenomes(fA, fB)

}
