% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(fitA, fitB)
}
\arguments{
\item{intercept}{}
}
\value{
result
}
\description{

}
\details{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan(rate=0, fp_error_rate=1, fn_error_rate=1)
simB <- simulate_pan(rate=0, fp_error_rate=5, fn_error_rate=5)
fA <- panstripe(simA$pa, simA$tree, nboot=0)
fB <- panstripe(simB$pa, simB$tree, nboot=0)
comp <- compare_pangenomes(fA, fB)

}
