% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pangenomes.R
\name{compare_pangenomes}
\alias{compare_pangenomes}
\title{compare_pangenomes}
\usage{
compare_pangenomes(resA, resB, ...)
}
\arguments{
\item{fitB}{}
}
\value{
result
}
\description{

}
\details{
Compares pangenomes using anova
}
\examples{

simA <- simulate_pan()
simB <- simulate_pan(rate=1e-3)
resA <- fit_tweedie(simA$pa, simA$tree)
resB <- fit_tweedie(simB$pa, simB$tree)
comp <- compare_pangenomes(resA, resB)

}
