% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tweedie.R
\name{fit_tweedie}
\alias{fit_tweedie}
\title{fit_tweedie}
\usage{
fit_tweedie(pa, tree, nboot = 100)
}
\arguments{
\item{pa}{a binary gene presence/absence matrix with genes as columns and genomes as rows. The rownames should match the tip.labels of the corresponding phylogeny.}

\item{tree}{a core gene phylogney of class \link{phylo}}

\item{nboot}{the number of bootstrap replicates to perform (default=100)}
}
\value{
result
}
\description{

}
\details{
Fits a tweedie...
}
\examples{

sim <- simulate_pan()
pa <- sim$pa
tree <- sim$tree
nboot <- 100
res <- fit_tweedie(sim$pa, sim$tree)


}
