% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panstripe.R
\name{panstripe}
\alias{panstripe}
\title{panstripe}
\usage{
panstripe(
  pa,
  tree,
  asr_method = "max.parsimony",
  min_depth = NULL,
  family = "Tweedie",
  ci_type = "norm",
  nboot = 1000,
  boot_type = "branch",
  conf = 0.95,
  boot_pvalue = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{pa}{a binary gene presence/absence matrix with genes as columns and genomes as rows. The rownames should match the tip.labels of the corresponding phylogeny.}

\item{tree}{a core gene phylogeny of class \link{phylo}}

\item{asr_method}{method used to perform ancestral state reconstruction. Can be either 'max_parsimony' (default) or 'ML' (experimental)}

\item{min_depth}{the minimum depth of a branch to be included in the regression. All branches are included by default.}

\item{family}{the family used by glm. One of 'Tweedie', 'Poisson', 'Gamma' or 'Gaussian'. (default='Tweedie')}

\item{ci_type}{the method used to calculate the bootstrap CI (default='bca'). See \link[boot]{boot.ci} for more details.}

\item{nboot}{the number of bootstrap replicates to perform (default=100)}

\item{boot_type}{whether to resample by 'branch', the default, or by 'gene'}

\item{conf}{A scalar indicating the confidence level of the required intervals (default=0.95).}

\item{boot_pvalue}{whether or not to calculate bootstrap p-values (default=FALSE)}

\item{quiet}{whether to print progress information (default=FALSE)}
}
\value{
a panfit object with the resulting parameter estimates and bootstrap replicates
}
\description{
Fits a Tweedie distribution to the inferred gene gain and loss events along each branch of a given phylogeny. 
Includes covariates to control for the impact of annotation errors and the depth of ancestral branches.
}
\examples{

sim <- simulate_pan(rate=5e-4, mean_trans_size=3, fn_error_rate=2, fp_error_rate=2)
pa <- sim$pa
tree <- sim$tree
nboot <- 100
family <- 'Tweedie'
ci_type='perc'
boot_type='branch'
conf=0.95
res <- panstripe(sim$pa, sim$tree, nboot=100, family='Tweedie')
res$summary

}
