% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_params.R
\name{plot_dist_params}
\alias{plot_dist_params}
\title{plot_dist_params}
\usage{
plot_dist_params(res, plot = TRUE)
}
\arguments{
\item{res}{the result of running the `fit_tweedie` function. Multiple results can be passed as a named list.}

\item{plot}{(default=FALSE)}
}
\value{
result
}
\description{

}
\details{
Plots the fitted pangenome tweedie regression model.
}
\examples{

sim <- simulate_pan(rate=1e-3)
res <- fit_tweedie(sim$pa, sim$tree)
plot_dist_params(res)

}
