% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_params.R
\name{plot_dist_params}
\alias{plot_dist_params}
\title{plot_dist_params}
\usage{
plot_dist_params(
  fit,
  plot = TRUE,
  text_size = 14,
  color_pallete = 6,
  boot_ci = FALSE
)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple results can be passed as a named list.}

\item{plot}{(default=FALSE)}

\item{text_size}{the base text size of the plot (default=14)}

\item{color_pallete}{the pallete number passed to `scale_fill_brewer`}

\item{boot_ci}{whether to use bootstrap CI or standard error}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the fitted pangenome tweedie regression model.
}
\examples{

sim <- simulate_pan(rate=1e-3, mean_trans_size = 1)
fit <- panstripe(sim$pa, sim$tree, nboot=100, ci_type='perc')
plot_dist_params(fit)
simB <- simulate_pan(rate=1e-3, mean_trans_size = 10)
fitB <- panstripe(simB$pa, simB$tree, nboot=100, ci_type='perc')
plot_dist_params(list(a=fit,b=fitB))

}
