% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gain_loss.R
\name{plot_gain_loss}
\alias{plot_gain_loss}
\title{plot_gain_loss}
\usage{
plot_gain_loss(fit, text_size = 14, color_pallete = 7)
}
\arguments{
\item{fit}{the output of running 'panstripe'}

\item{text_size}{adjusts the size of text in the plot}

\item{color_pallete}{the colour pallete to use. A number between 1 & 9. See 'scale_colour_brewer' for more details}
}
\value{
a plot of the phylogeny coloured by the inferred total gene gain/loss events per branch
}
\description{

}
\details{
Plots a expected gain and loss events onto the branches of the phylogeny.
}
\examples{

sim <- simulate_pan(rate=1e-3)
fA <- panstripe(sim$pa, sim$tree, nboot=0)
plot_gain_loss(fA, color_pallete=7)

}
