% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pangenome_fits.R
\name{plot_pangenome_fits}
\alias{plot_pangenome_fits}
\title{plot_pangenome_fits}
\usage{
plot_pangenome_fits(
  fit,
  type = "cumulative",
  ci = TRUE,
  plot = TRUE,
  legend = TRUE,
  text_size = 14,
  color_pallete = 6
)
}
\arguments{
\item{fit}{the result of running the `panstripe` function. Multiple fits can be passed as a named list.}

\item{type}{either 'cumulative' (default) or 'individual'. Indicates whether or not to plot the cumulative sum of core and accessory branch lengths.}

\item{ci}{whether or not to include the confidence interval ribbon (default=TRUE)}

\item{plot}{whether to generate the plot (default) or return a data.frame}

\item{legend}{toggles the display of the legend on and off}

\item{text_size}{the base text size of the plot (default=14)}

\item{color_pallete}{the pallete number passed to `scale_fill_brewer`}
}
\value{
either a ggplot2 object or a `data.frame` with the data needed to recreate the plot
}
\description{
Plots the fitted pangenome tweedie regression model.
}
\examples{

sim <- simulate_pan(rate=0)
fA <- panstripe(sim$pa, sim$tree, nboot=10, ci_type='perc')
plot_pangenome_fits(fA, color_pallete=6)
sim <- simulate_pan(rate=1e-2)
fB <- panstripe(sim$pa, sim$tree, nboot=10, ci_type='perc')
plot_pangenome_fits(list(a=fA,b=fB), color_pallete=6, ci=TRUE, type='individual')
plot_pangenome_fits(list(a=fA,b=fB), color_pallete=6, ci=TRUE, type='cumulative')

}
