% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfreemap.R
\name{sfreemap}
\alias{sfreemap}
\title{sfreemap}
\usage{
sfreemap(tree, pa, model = "ER", quiet = FALSE)
}
\arguments{
\item{tree}{tree}

\item{pa}{gene presence/absence matrix}

\item{model}{model}

\item{quiet}{turn off messages (FALSE)}
}
\value{
result
}
\description{
This function has been adapted and simplified from the code in the sfreemap R package (https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-017-1554-7).
}
\details{
Simulation free stochastic character mapping on a phylogenetic tree.
}
\examples{
\dontrun{
tree <- ape::rtree(10)
pa <- matrix(rbinom(10, 1, 0.5), ncol=1)
rownames(pa) <- tree$tip.label
sfreemap(tree, pa)
}

}
