% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfreemap.R
\name{sfreemap}
\alias{sfreemap}
\title{sfreemap}
\usage{
sfreemap(tree, pa, type = "standard", model = "ER", quiet = FALSE)
}
\arguments{
\item{tree}{}

\item{pa}{}

\item{type}{}

\item{model}{}

\item{method}{}

\item{tol}{}
}
\value{
result
}
\description{
This function has been adapted and simplified from the code in the sfreemap R package.
}
\details{
Simulation free stochastic character mapping on a phylogenetic tree.
}
\examples{

tree <- ape::rtree(10)
pa <- matrix(rbinom(10, 1, 0.5), ncol=1)
rownames(pa) <- tree$tip.label
sfreemap(tree, pa)


}
